+++ app/forgot-password/types.ts
@@ -1,0 +1,9 @@
+export type ForgotPasswordState = {
+  error: string | null;
+  submitted: boolean;
+};
+
+export const INITIAL_FORGOT_PASSWORD_STATE: ForgotPasswordState = {
+  error: null,
+  submitted: false,
+};
